refactor(sprint-1/a1): delete dead leaf files (discord, local-stt, cognitive forwarder)
Karpathy refactor sprint 1, PR-A1 — pure deletes, zero behavior change.

- src/gateway/voice/adapters/local-stt.ts (0 LOC empty file)
- src/gateway/channels/discord.ts (54 LOC adapter, zero operator usage —
  Telegram is the sole active channel; if Discord is ever needed it'll
  be reimplemented cleanly without legacy cruft)
- src/infra/cli/utils/cognitive.ts (5 LOC re-export forwarder — three
  callers updated to import directly from src/cognitive/runtime-support)
- discord.js dependency dropped from package.json (^13.17.1 + transitive
  deps no longer pulled by anything)
- tests/unit/gateway-channels.test.ts: discord section removed,
  telegram + splitText sections kept

Plan reference: ~/.claude/plans/glowing-knitting-lobster.md sprint 1.
