+++ package.json
@@ -192,7 +192,6 @@
     "chalk": "^5.6.2",
     "cli-progress": "^3.12.0",
     "commander": "^14.0.3",
-    "discord.js": "^13.17.1",
     "dotenv": "^17.3.1",
     "fastify": "^5.8.2",
     "grammy": "^1.41.1",
