--- src/infra/cli/utils/cognitive.ts
@@ -1,5 +1,0 @@
-export {
-  DEFAULT_COGNITIVE_CHAIN_LIMITS,
-  loadCognitiveBlocks,
-  normalizeCognitiveBlock,
-} from '../../../cognitive/runtime-support.js';
